+++ src/main/java/gift/product/domain/ProductResponse.java
@@ -3,10 +3,10 @@ package gift.product.domain;
 public class ProductResponse {
     private Long id;
     private String name;
-    private Double price;
+    private Long price;
     private String imageUrl;
 
-    public ProductResponse(Long id, String name, Double price, String imageUrl) {
+    public ProductResponse(Long id, String name, Long price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -31,7 +31,7 @@ public class ProductResponse {
         return name;
     }
 
-    public Double getPrice() {
+    public Long getPrice() {
         return price;
     }
