+++ src/main/resources/db/schema.sql
@@ -32,7 +32,7 @@ CREATE TABLE product
 (
     product_id  BIGINT AUTO_INCREMENT PRIMARY KEY,
     name        VARCHAR(255) NOT NULL,
-    price DOUBLE NOT NULL,
+    price       BIGINT       NOT NULL,
     image_url   VARCHAR(255) NOT NULL,
     category_id BIGINT       NOT NULL,
     created_at  TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
